+++ server/modules/filter/cache/lrustoragemt.hh
@@ -21,7 +21,7 @@ class LRUStorageMT : public LRUStorage
 public:
     ~LRUStorageMT();
 
-    static LRUStorageMT* create(Storage* pstorage, size_t max_count, size_t max_size);
+    static LRUStorageMT* create(Storage* pstorage, uint64_t max_count, uint64_t max_size);
 
     cache_result_t get_info(uint32_t what,
                             json_t** ppInfo) const;
@@ -36,7 +36,7 @@ public:
     cache_result_t del_value(const CACHE_KEY& key);
 
 private:
-    LRUStorageMT(Storage* pstorage, size_t max_count, size_t max_size);
+    LRUStorageMT(Storage* pstorage, uint64_t max_count, uint64_t max_size);
 
     LRUStorageMT(const LRUStorageMT&);
     LRUStorageMT& operator = (const LRUStorageMT&);
